requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
demucs==4.0.1
essentia==2.1b6.dev1110
midiutil==1.2.1
matplotlib==3.8.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
supabase==2.0.2
modal==0.63.0
postgrest==0.14.0